requirements.txt
requests==2.28.1
biopython==1.81
python-dotenv==0.21.0
